--- src/main/java/konkuk/thip/feed/application/port/in/FeedShowMineUseCase.java
@@ -1,8 +1,0 @@
-package konkuk.thip.feed.application.port.in;
-
-import konkuk.thip.feed.adapter.in.web.response.FeedShowMineResponse;
-
-public interface FeedShowMineUseCase {
-
-    FeedShowMineResponse showMyFeeds(Long userId, String cursor);
-}
